changed scene structure
